Working towards getting the batches ready. So far: - Imported most of the code from the old method - It's returning the right values
Problem to deal with is the way it handles paging - do we want to make all the calls together, or shall we use the paging mechanism because it's there? Ideally we just use the standard method. Perhaps there could be a flag to ask whether we should look for paging or not.